simplify the unpacker result models
